requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.39
botocore==1.34.39
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
jmespath==1.0.1
multidict==6.0.5
Pint==0.23
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
websockets==12.0
werk24==1.9.0
yarl==1.9.4
